Move JWT generation into a seperate generator Add JWT security variables into container class
