+++ src/main/java/com/tungstun/security/util/jwt/JwtCredentials.java
@@ -1,0 +1,13 @@
+package com.tungstun.security.util.jwt;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Component
+public class JwtCredentials {
+    @Value("${security.jwt.secret}")
+    protected String jwtSecret;
+
+    @Value("${security.jwt.expiration-in-ms}")
+    protected Integer jwtExpirationInMs;
+}
